split cruft into modules
